+++ src/Builders/BrandBuilder.php
@@ -1,26 +1,26 @@
 <?php
-declare(strict_types = 1);
+
 namespace App\Builders;
 
-use BBC\ProgrammesPagesService\Domain\Entity\Episode;
+use BBC\ProgrammesPagesService\Domain\Entity\Brand;
 use BBC\ProgrammesPagesService\Domain\Entity\Options;
 use BBC\ProgrammesPagesService\Domain\ValueObject\Pid;
 use BBC\ProgrammesPagesService\Domain\ValueObject\Synopses;
 
-class EpisodeBuilder extends AbstractProgrammeItemBuilder implements BuilderInterface
+class BrandBuilder extends AbstractProgrammeContainerBuilder implements BuilderInterface
 {
-    /** @var int */
-    private $aggregatedBroadcastsCount;
-
-    /** @var int */
-    private $availableClipsCount;
-
     private function __construct()
     {
+        $this->aggregatedBroadcastsCount =2;
+        $this->aggregatedEpisodesCount = 4;
+        $this->availableClipsCount = 3;
+        $this->availableEpisodesCount = 2;
+        $this->isPodcastable = true;
+        $this->expectedChildCount = 2;
         $this->dbAncestryIds = [1212];
         $this->pid = new Pid('d00744wz');
-        $this->title = 'my episode title';
-        $this->searchTitle = 'my search episode title';
+        $this->title = 'My brand title';
+        $this->searchTitle = 'My search title';
         $this->synopses = new Synopses('My short synopsis', 'my a very medium no too much synopsis', 'my extremely boring and endless text for my long synopsis');
         $this->image = ImageBuilder::default()->build();
         $this->promotionsCount = 10;
@@ -29,35 +29,26 @@ class EpisodeBuilder extends AbstractProgrammeItemBuilder implements BuilderInte
         $this->isStreamable = true;
         $this->isStreamableAlternate = true;
         $this->contributionsCount = 2;
-        $this->mediaType = 'audio';
-        $this->segmentEventCount = 19;
-        $this->aggregatedGalleriesCount = 2;
+        $this->aggregatedGalleriesCount = 3;
         $this->options = new Options();
-        $this->aggregatedBroadcastsCount = 10;
-        $this->availableClipsCount = 14;
-        $this->duration = 6400;
+        $this->aggregatedGalleriesCount = 6;
     }
 
-    public function withAggregatedBroadcastsCount(int $aggregatedBroadcastsCount)
-    {
-        $this->aggregatedBroadcastsCount = $aggregatedBroadcastsCount;
-        return $this;
-    }
-
-    public function withAvailableClipsCount(int $availableClipsCount)
-    {
-        $this->availableClipsCount = $availableClipsCount;
-        return $this;
-    }
-
+    /**
+     * Create a builder with default state.
+     * This default state reduce the amount of steps necessary when trying to get a final entity in an specified state
+     */
     public static function default()
     {
         return new self();
     }
 
-    public function build(): Episode
+    /**
+     * Given that we have the desired state, then create the real entity we want
+     */
+    public function build(): Brand
     {
-        return new Episode(
+        return new Brand(
             $this->dbAncestryIds,
             $this->pid,
             $this->title,
@@ -70,11 +61,12 @@ class EpisodeBuilder extends AbstractProgrammeItemBuilder implements BuilderInte
             $this->isStreamable,
             $this->isStreamableAlternate,
             $this->contributionsCount,
-            $this->mediaType,
-            $this->segmentEventCount,
             $this->aggregatedBroadcastsCount,
+            $this->aggregatedEpisodesCount,
             $this->availableClipsCount,
+            $this->availableEpisodesCount,
             $this->aggregatedGalleriesCount,
+            $this->isPodcastable,
             $this->options,
             // optionals
             $this->parent,
@@ -83,10 +75,7 @@ class EpisodeBuilder extends AbstractProgrammeItemBuilder implements BuilderInte
             $this->genres,
             $this->formats,
             $this->firstBroadcastDate,
-            $this->releaseDate,
-            $this->duration,
-            $this->streamableFrom,
-            $this->streamableUntil
+            $this->expectedChildCount
         );
     }
 }
